% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gmdh.R
\name{predict.gmdh}
\alias{predict.gmdh}
\title{Calculates GMDH model predictions for new data.}
\usage{
\method{predict}{gmdh}(object, data, ...)
}
\arguments{
\item{object}{GMDH model - an object of class 'gmdh'}

\item{data}{a data matrix for which the predictions are calculated}

\item{...}{other arguments passed to this function}
}
\value{
a vector of predictions generated by the model
}
\description{
Calculates GMDH model predictions for new data.
}

