% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gmdh.R
\docType{package}
\name{rgmdh}
\alias{rgmdh}
\alias{rgmdh-package}
\title{rgmdh: an implementation of the Group Method of Data Handling regression method}
\description{
An implementation of the Group Method of Data Handling proposed by A.G. Ivakhnenko in 1966.  The package allows solving regression tasks by creating high degree polynomials thus only numerical attributes are allowed.
}

