% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gmdh.R
\name{train}
\alias{train}
\title{Trains a GMDH model which solves the y=f(x) regression task.}
\usage{
train(x, y, M = ncol(y), val.size = 0.25)
}
\arguments{
\item{x}{a matrix with independent (input) variables}

\item{y}{a numeric vector with the dependent (output) variable}

\item{M}{the number of polynomials which get selected at the end of i-th iteration, and are processed durint iteration i+1}

\item{val.size}{fraction of the available data used for validation}
}
\value{
a trained GMDH model - an object of class 'gmdh'
}
\description{
Trains a GMDH model which solves the y=f(x) regression task.
}

